Centralize repeated logic
