+++ client/my-sites/customer-home/cards/launchpad/index.tsx
@@ -1,27 +1,34 @@
 import { Button, CircularProgressBar, Gridicon } from '@automattic/components';
-import { updateLaunchpadSettings, useLaunchpad } from '@automattic/data-stores';
-import { Launchpad, Task } from '@automattic/launchpad';
+import { updateLaunchpadSettings, useLaunchpad, SiteDetails } from '@automattic/data-stores';
+import { Launchpad, Task, setUpActionsForTasks } from '@automattic/launchpad';
 import { useTranslate } from 'i18n-calypso';
 import { useEffect, useState } from 'react';
 import { recordTracksEvent } from 'calypso/lib/analytics/tracks';
 import { useSelector } from 'calypso/state';
-import { getSiteSlug } from 'calypso/state/sites/selectors';
+import { getSite, getSiteSlug } from 'calypso/state/sites/selectors';
 import { getSelectedSiteId } from 'calypso/state/ui/selectors';
+import ShareSiteModal from '../../components/share-site-modal';
 import type { AppState } from 'calypso/types';
 
 import './style.scss';
 
 interface CustomerHomeLaunchpadProps {
 	checklistSlug: string;
-	taskFilter: ( tasks: Task[] ) => Task[];
+	extraActions?: any;
 }
 
 const CustomerHomeLaunchpad = ( {
 	checklistSlug,
-	taskFilter,
+	extraActions = {},
 }: CustomerHomeLaunchpadProps ): JSX.Element => {
+	const launchpadContext = 'customer-home';
 	const siteId = useSelector( getSelectedSiteId );
 	const siteSlug = useSelector( ( state: AppState ) => getSiteSlug( state, siteId ) );
+
+	// The type definition for getSite is incorrect, it returns a SiteDetails object
+	const site = useSelector(
+		( state: AppState ) => siteId && ( getSite( state as object, siteId ) as any as SiteDetails )
+	);
 
 	const translate = useTranslate();
 	const [ isDismissed, setIsDismissed ] = useState( false );
@@ -33,9 +40,26 @@ const CustomerHomeLaunchpad = ( {
 		setIsDismissed( initialIsChecklistDismissed );
 	}, [ initialIsChecklistDismissed ] );
 
+	const [ shareSiteModalIsOpen, setShareSiteModalIsOpen ] = useState( false );
+
 	const numberOfSteps = checklist?.length || 0;
 	const completedSteps = ( checklist?.filter( ( task: Task ) => task.completed ) || [] ).length;
 	const tasklistCompleted = completedSteps === numberOfSteps;
+	const tracksData = { recordTracksEvent, checklistSlug, tasklistCompleted, launchpadContext };
+
+	const defaultExtraActions = {
+		setShareSiteModalIsOpen,
+		...extraActions,
+	};
+
+	const taskFilter = ( tasks: Task[] ) => {
+		return setUpActionsForTasks( {
+			tasks,
+			siteSlug,
+			tracksData,
+			extraActions: defaultExtraActions,
+		} );
+	};
 
 	useEffect( () => {
 		// Record task list view as a whole.
@@ -109,6 +133,9 @@ const CustomerHomeLaunchpad = ( {
 					</div>
 				) }
 			</div>
+			{ shareSiteModalIsOpen && site && (
+				<ShareSiteModal setModalIsOpen={ setShareSiteModalIsOpen } site={ site } />
+			) }
 			<Launchpad siteSlug={ siteSlug } checklistSlug={ checklistSlug } taskFilter={ taskFilter } />
 		</div>
 	);
