+++ client/my-sites/customer-home/cards/launchpad/intent-newsletter.tsx
@@ -1,46 +1,9 @@
-import { useLaunchpad } from '@automattic/data-stores';
-import { setUpActionsForTasks } from '@automattic/launchpad';
-import { useState } from 'react';
-import { recordTracksEvent } from 'calypso/lib/analytics/tracks';
-import { useSelector } from 'calypso/state';
-import { getSite, getSiteSlug } from 'calypso/state/sites/selectors';
-import { getSelectedSiteId } from 'calypso/state/ui/selectors';
-import ShareSiteModal from '../../components/share-site-modal';
 import CustomerHomeLaunchpad from '.';
-import type { Task } from '@automattic/launchpad';
-import type { AppState } from 'calypso/types';
 
 const LaunchpadIntentNewsletter = ( { checklistSlug }: { checklistSlug: string } ): JSX.Element => {
-	const siteId = useSelector( getSelectedSiteId );
-	const siteSlug = useSelector( ( state: AppState ) => getSiteSlug( state, siteId ) );
-	const site = useSelector( ( state: AppState ) => siteId && getSite( state, siteId ) );
-	const launchpadContext = 'customer-home';
-
-	const {
-		data: { checklist },
-	} = useLaunchpad( siteSlug, checklistSlug );
-
-	const [ shareSiteModalIsOpen, setShareSiteModalIsOpen ] = useState( false );
-
-	const numberOfSteps = checklist?.length || 0;
-	const completedSteps = ( checklist?.filter( ( task: Task ) => task.completed ) || [] ).length;
-	const tasklistCompleted = numberOfSteps > 0 && completedSteps === numberOfSteps;
-	const tracksData = { recordTracksEvent, checklistSlug, tasklistCompleted, launchpadContext };
-	const extraActions = { setShareSiteModalIsOpen };
-
-	const sortedTasksWithActions = ( tasks: Task[] ) => {
-		return setUpActionsForTasks( { tasks, siteSlug, tracksData, extraActions } );
-	};
-
 	return (
 		<>
-			<CustomerHomeLaunchpad
-				checklistSlug={ checklistSlug }
-				taskFilter={ sortedTasksWithActions }
-			></CustomerHomeLaunchpad>
-			{ shareSiteModalIsOpen && site && (
-				<ShareSiteModal setModalIsOpen={ setShareSiteModalIsOpen } site={ site } />
-			) }
+			<CustomerHomeLaunchpad checklistSlug={ checklistSlug }></CustomerHomeLaunchpad>
 		</>
 	);
 };
